# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Playlisted API
  version: 0.2.0
  summary: OpenAPI-first REST API for the Playlisted audio platform.
  description: |
    Contract for the Playlisted backend. The server validates requests and responses
    against this document, and the client SDK is generated from the same source.
servers:
  - url: http://localhost:4000
    description: Local development
security: []
...
paths:
...
  /api/v1/me/recordings:
    get:
      tags: [Me]
      operationId: listMyRecordings
      summary: List recordings uploaded by the current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Owned recordings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/favorites/recordings:
    get:
      tags: [Me]
      operationId: listFavoriteRecordings
      summary: List the current user's favorited recordings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Favorite recordings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteRecordingsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/favorites/recordings/{recordingId}:
    post:
      tags: [Me]
      operationId: addFavoriteRecording
      summary: Add a recording to favorites
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: recordingId
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteSavedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recording not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Me]
      operationId: removeFavoriteRecording
      summary: Remove a recording from favorites
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: recordingId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/me/analytics/recordings:
    get:
      tags: [Analytics]
      operationId: getAnalyticsRecordings
      summary: Get per-track analytics for the current creator
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: range
          schema:
            $ref: '#/components/schemas/ChartRange'
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [plays, duration, completion]
            default: plays
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Per-track analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsRecordingsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/recordings:
    get:
      tags: [Recordings]
      operationId: listRecordings
      summary: List recordings
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: uploaderId
          schema:
            type: string
        - in: query
          name: playlistId
          schema:
            type: string
        - in: query
          name: recordingType
          schema:
            $ref: '#/components/schemas/RecordingType'
        - in: query
          name: visibility
          schema:
            $ref: '#/components/schemas/Visibility'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/PublishStatus'
      responses:
        '200':
          description: Recording collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingListResponse'
    post:
      tags: [Recordings]
      operationId: createRecording
      summary: Create a recording and attach it to its canonical playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordingRequest'
      responses:
        '201':
          description: Recording created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingDetail'
        '404':
          description: Uploader or playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Recording conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/recordings/{recordingId}:
    get:
      tags: [Recordings]
      operationId: getRecordingById
      summary: Get recording detail
      parameters:
        - in: path
          name: recordingId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recording detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingDetail'
        '404':
          description: Recording not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    UserRole:
      type: string
      enum: [LISTENER, CREATOR, EDITOR, ADMIN]
...
    PlaylistType:
      type: string
      enum: [PLAYLIST, ALBUM, MIX, PODCAST_CHANNEL, RELEASE]
    RecordingType:
      type: string
      enum: [SONG, PODCAST_EPISODE, MIX_SEGMENT, OTHER]
    Visibility:
      type: string
      enum: [PUBLIC, UNLISTED, PRIVATE]
    PublishStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details: {}
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    OwnerSummary:
      type: object
      additionalProperties: false
      required: [id, username, displayName, role]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
    RecordingSummary:
      type: object
      additionalProperties: false
      required:
        [id, uploaderId, publishedPlaylistId, title, audioUrl, recordingType, visibility, status, explicit, playCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        uploaderId:
          type: string
        publishedPlaylistId:
          type: string
        title:
          type: string
        description:
          type: [string, 'null']
        audioUrl:
          type: string
          format: uri-reference
        audioMimeType:
          type: [string, 'null']
        audioBytes:
          type: [integer, 'null']
        durationSeconds:
          type: [integer, 'null']
        artworkUrl:
          type: [string, 'null']
          format: uri-reference
        recordingType:
          $ref: '#/components/schemas/RecordingType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        trackNumber:
          type: [integer, 'null']
        episodeNumber:
          type: [integer, 'null']
        explicit:
          type: boolean
        releaseDate:
          type: [string, 'null']
          format: date-time
        publishedAt:
          type: [string, 'null']
          format: date-time
        playCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PublishedPlaylistReference:
      type: object
      additionalProperties: false
      required: [id, ownerId, title, slug, type, visibility, status]
      properties:
        id:
          type: string
        ownerId:
          type: string
        title:
          type: string
        slug:
          type: string
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
    RecordingDetail:
      allOf:
        - $ref: '#/components/schemas/RecordingSummary'
        - type: object
          additionalProperties: false
          required: [uploader, publishedPlaylist]
          properties:
            uploader:
              $ref: '#/components/schemas/OwnerSummary'
            publishedPlaylist:
              $ref: '#/components/schemas/PublishedPlaylistReference'
    RecordingListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RecordingSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateRecordingRequest:
      type: object
      additionalProperties: false
      required: [uploaderId, publishedPlaylistId, title, audioUrl]
      properties:
        uploaderId:
          type: string
        publishedPlaylistId:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 191
        description:
          type: [string, 'null']
        audioUrl:
          type: string
          format: uri-reference
        audioMimeType:
          type: [string, 'null']
        audioBytes:
          type: [integer, 'null']
          minimum: 0
        durationSeconds:
          type: [integer, 'null']
          minimum: 0
        artworkUrl:
          type: [string, 'null']
          format: uri-reference
        recordingType:
          $ref: '#/components/schemas/RecordingType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        trackNumber:
          type: [integer, 'null']
          minimum: 0
        episodeNumber:
          type: [integer, 'null']
          minimum: 0
        explicit:
          type: boolean
          default: false
        releaseDate:
          type: [string, 'null']
          format: date-time
        publishedAt:
          type: [string, 'null']
          format: date-time
...
    ChartRange:
      type: string
      enum: [7d, 30d, all]
      default: 30d
    ChartArtistRef:
      type: object
      additionalProperties: false
      required: [id, username, displayName, role]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
...
    AnalyticsRecordingItem:
      type: object
      additionalProperties: false
      required: [recordingId, title, totalPlays, totalPlaySeconds, completionRate]
      properties:
        recordingId:
          type: string
        title:
          type: string
        artworkUrl:
          type: [string, 'null']
          format: uri-reference
        durationSeconds:
          type: [integer, 'null']
        totalPlays:
          type: integer
        totalPlaySeconds:
          type: integer
        completionRate:
          type: number
    AnalyticsRecordingsResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsRecordingItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PersonalTrackItem:
      type: object
      required: [id, uploaderId, publishedPlaylistId, title, audioUrl, recordingType, visibility, status, explicit, playCount, createdAt, updatedAt, uploader]
      properties:
        id:
          type: string
        uploaderId:
          type: string
        publishedPlaylistId:
          type: string
        title:
          type: string
        description:
          type: [string, 'null']
        audioUrl:
          type: string
          format: uri-reference
        audioMimeType:
          type: [string, 'null']
        audioBytes:
          type: [integer, 'null']
        durationSeconds:
          type: [integer, 'null']
        artworkUrl:
          type: [string, 'null']
          format: uri-reference
        recordingType:
          $ref: '#/components/schemas/RecordingType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        trackNumber:
          type: [integer, 'null']
        episodeNumber:
          type: [integer, 'null']
        explicit:
          type: boolean
        releaseDate:
          type: [string, 'null']
          format: date-time
        publishedAt:
          type: [string, 'null']
          format: date-time
        playCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        uploader:
          $ref: '#/components/schemas/ChartArtistRef'
    FavoriteRecordingItem:
      allOf:
        - $ref: '#/components/schemas/PersonalTrackItem'
        - type: object
          required: [savedAt]
          properties:
            savedAt:
              type: string
              format: date-time
    FavoriteRecordingsResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteRecordingItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    FavoriteSavedResponse:
      type: object
      additionalProperties: false
      required: [id, recordingId, savedAt]
      properties:
        id:
          type: string
        recordingId:
          type: string
        savedAt:
          type: string
          format: date-time
...